sudah bisa read dan update data 2 byte

+++ src/EEPROMWearLevelWOcontrol.h
@@ -18,7 +18,7 @@ class MSN_EEPROMWearLevel{
        uint16_t address_toWrite;
     };
     void begin(MSN_EEPROMwlAddr* address[], uint16_t partition_length);
-    uint16_t read(uint16_t var_n, MSN_EEPROMwlAddr* address[]);
+    uint32_t read(uint16_t var_n, MSN_EEPROMwlAddr* address[]);
     void update(uint16_t var_n, MSN_EEPROMwlAddr* address[], uint8_t val);
     void update(uint16_t var_n, MSN_EEPROMwlAddr* address[], uint16_t val);
     void update(uint16_t var_n, MSN_EEPROMwlAddr* address[], uint32_t val);
